# api/swagger.yaml
...
swagger: "2.0"
info:
  description: "This is the HLL App API"
  version: "2.0.0"
  title: "Hyreslandslaget - HLL API"
host: "api.hyreslandslaget.se"
basePath: "/"
...
schemes:
- "https"
paths:
...
  /search:
    get:
      tags:
      - "Search"
      summary: "Search"
      description: ""
      operationId: "searchGET"
      produces:
      - "application/json"
      parameters:
      - name: "query"
        in: "query"
        description: "The query"
        required: true
        type: "string"
        x-example: "Truck"
      - name: "type"
        in: "query"
        description: "The type of domain to where search should be perform."
        required: false
        type: "string"
        enum:
        - "equipment"
        - "category"
        - "depot"
        x-example: "equipment"
      - name: "limit"
        in: "query"
        description: "The limit for search results."
        required: false
        type: "string"
        x-example: "10"
      - name: "offset"
        in: "query"
        description: "The offset for search results."
        required: false
        type: "string"
        x-example: "10"
      responses:
        200:
          description: "OK"
          examples:
            application/json:
              items:
              - type: "equipment"
                displayName: "Machine"
                nextOffset: 100
                results:
                - reference: "21234"
                  displayText: "DepotTruck"
              - reference: "93212345"
                displayText: "DepotTruck"
                sraCode: "93212345"
          headers: {}
          schema:
            type: "object"
            required:
            - "items"
            properties:
              items:
                type: "array"
                description: "The different search result sections, Equipment, Depot\
                  \ etc"
                items: {}
        403:
          description: "Forbidden"
          examples:
            application/json:
              message: "Resource not allowed"
              code: 403
              errors:
              - domain: "global"
                reason: "backendError"
                message: "Error message"
          headers: {}
          schema:
            type: "object"
            required:
            - "code"
            - "message"
            properties:
              message:
                type: "string"
                description: "The meessage of Error."
              code:
                type: "number"
                description: "The error code."
              errors:
                type: "array"
                description: "Array of errors."
                items: {}
        404:
          description: "Not Found"
          examples:
            application/json:
              message: "Resource not found"
              code: 404
              errors:
              - domain: "global"
                reason: "backendError"
                message: "Error message"
          headers: {}
          schema:
            type: "object"
            required:
            - "code"
            - "message"
            properties:
              message:
                type: "string"
                description: "The meessage of Error."
              code:
                type: "number"
                description: "The error code."
              errors:
                type: "array"
                description: "Array of errors."
                items: {}
      x-swagger-router-controller: "Search"
...